cdk/pyproject.toml
[tool.poetry]
name = "eval-sandbox"
version = "0.1.0"
description = "AWS CDK project managed by Poetry"
authors = [
    "ShiYue <yue.shi@amptalk.co.jp>"
]
readme = "README.md"
packages = [
    {include = "stacks"},
    {include = "app.py"}
    ]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
aws-cdk-lib = "^2.181.1"
constructs = "^10.4.2"
yamllint = "^1.35.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
pylint = "^3.1.0"
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pylint.messages_control]
disable = ["C0111", "C0103", "C0330", "C0326"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3
